refactoring auth and seperating middlewares
